:hammer: Moved getStyle functions.
