+++ src/profile-card/get-styles.js
@@ -1,0 +1,17 @@
+import styles from './styles';
+import combineStyles from '../internal/combine-styles';
+import colors from '../settings/colors';
+
+function root(themeColor, avatar, overrideStyle) {
+  const color = themeColor || colors.theme;
+
+  const style = { ...styles.profileCard, backgroundColor: color };
+
+  if (avatar) {
+    return combineStyles(combineStyles(style, styles.avatar), overrideStyle);
+  }
+
+  return combineStyles(style, overrideStyle);
+}
+
+export default { root };
